AS-2 feat: add Minio configuration and client setup; refactor StorageException constructor
