+++ src/main/java/lk/gov/mohe/adminsystem/storage/MinioConfig.java
@@ -1,0 +1,26 @@
+package lk.gov.mohe.adminsystem.storage;
+
+import io.minio.MinioClient;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class MinioConfig {
+    @Value("${custom.minio.endpoint}")
+    private String endpoint;
+
+    @Value("${custom.minio.access-key}")
+    private String accessKey;
+
+    @Value("${custom.minio.secret-key}")
+    private String secretKey;
+
+    @Bean
+    public MinioClient minioClient() {
+        return MinioClient.builder()
+            .endpoint(endpoint)
+            .credentials(accessKey, secretKey)
+            .build();
+    }
+}
