+++ src/main/java/lk/gov/mohe/adminsystem/storage/StorageException.java
@@ -1,6 +1,10 @@
 package lk.gov.mohe.adminsystem.storage;
 
 public class StorageException extends RuntimeException {
+    public StorageException(String message) {
+        super(message);
+    }
+
     public StorageException(String message, Throwable cause) {
         super(message, cause);
     }
